requirements.txt
asyncpg
databases
fastapi==0.115.0
uvicorn[standard]==0.30.6

# PostgreSQL
psycopg2-binary==2.9.9

# RabbitMQ
aio-pika==9.4.1

# Data handling
pandas==2.2.3

# Environment variables
python-dotenv==1.0.1
requests
python-multipart
click